+++ CHANGELOG.md
@@ -1,7 +1,19 @@
 # Change Log
-All notable changes to the "inkton-nest" extension will be documented in this file.
+All notable changes to the "nester-develop" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
 ## [Unreleased]
 - Initial release
+
+## [0.1.1] - 2017-10-11
+### Changed
+- Changed the extension id used for menu items were from an earlier version.
+
+## [0.2.1] - 2017-10-11
+### Changed
+- Changed the progress messages to appear on an output channel.
+### Added
+- Added github project and issues links to package.json.
+- Added a new icon concept for trial.
+